API_REST_Unity/api_rest/requirements.txt
pymongo==4.6.1
fastapi==0.109.0
uvicorn==0.26.0
python-dotenv==1.0.1
graphene==3.3.0
graphene_pydantic==0.6.1
strawberry-graphql==0.219.2
mariadb==1.1.9